requirements.txt
python-json-logger==2.0.7
fastapi==0.95.1
uvicorn==0.21.1
requests==2.28.2
pydantic==1.10.7
PyYAML==5.4.1
jsonschema==4.17.3
httpx==0.24.0